+++ TP2_MVC_Twig/view/livre/livre-index.php
@@ -1,7 +1,10 @@
+{{ include('main-menu.php', {title: 'Livre', pageHeader: 'Livre liste'})}}
 <body>
     <main>
         <h1>Liste des livres en vente</h1>
-        <a href="livre/create" class="bouton">+ Ajouter un livre</a>
+        {% if session.lvlAccess >= 2 %}
+            <a href="livre/create" class="bouton">+ Ajouter un livre</a>
+        {% endif %}
         <table>
             <thead>
                 <tr>
@@ -19,8 +22,12 @@
                         <td>{{ livre.Description }}</td>
                         <td>{{ livre.nombrePages }}</td>
                         <td>{{ livre.prix }}</td>
-                        <td class='modify'><a href='livre/modifier/{{ livre.id }}'>Modifier</a></td>
-                        <td class='delete'><a href='livre/delete/{{ livre.id }}'>Supprimer</a></td>
+                        {% if session.lvlAccess >= 1 %}
+                            <td class='modify'><a href='livre/modifier/{{ livre.id }}'>Modifier</a></td>
+                        {% endif %}
+                        {% if session.lvlAccess >= 2 %}
+                            <td class='delete'><a href='livre/delete/{{ livre.id }}'>Supprimer</a></td>
+                        {% endif %}
                     </tr>
                     {% endfor %}
             {% endif %}
