+++ TP2_MVC_Twig/view/membre/membre-index.php
@@ -1,6 +1,9 @@
+{{ include('main-menu.php', {title: 'Membre', pageHeader: 'Membre liste'})}}
 <main>
     <h1>Membre</h1>
-    <a href="membre/create" class="bouton">+ Ajouter un membre</a>
+    {% if session.lvlAccess >= 2 %}
+        <a href="membre/create" class="bouton">+ Ajouter un membre</a>
+    {% endif %}
     <table>
         <thead>
             <tr>
@@ -18,8 +21,12 @@
                     <td>{{ membre.adresse }}</td>
                     <td>{{ membre.phone }}</td>
                     <td>{{ membre.courriel }}</td>
-                    <td class='modify'><a href='membre/modifier/{{membre.id}}'>Modifier</a></td>
-                    <td class='delete'><a href='membre/delete'>Supprimer</a></td>
+                    {% if session.lvlAccess >= 1 %}
+                        <td class='modify'><a href='membre/modifier/{{membre.id}}'>Modifier</a></td>
+                    {% endif %}
+                    {% if session.lvlAccess >= 2 %}
+                        <td class='delete'><a href='membre/delete'>Supprimer</a></td>
+                    {% endif %}
                 </tr>
                 {% endfor %}
             {% endif %}
